app/requirements.txt
protobuf
googleapis-common-protos
google-api-python-client
google-cloud-container